TST get_result error + RFC solvers test
